Datasets
The datasets referred to in the Video "Combining and Merging Datasets" can be found in the SAS file attached. Students can open this file in SAS Studio and browse the Libraries section, where they will find respective datasets in the 'LIBREF' library. 

For Concatenation: Please use these two datasets in LIBREF library -> "libref.lefttab" and "libref.righttab".

For Interleaving: Please use these two datasets in LIBREF library -> "libref.fashion" and "libref.electric".

For One-to-One Reading:  Run the commands in SAS file (from line #98) to create two temporary datasets "Customers" and "Sales".

For One-to-One Merging: Same datasets are used as used in the One-to-One Reading operation.